% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderEstimate.R
\name{as_feature}
\alias{as_feature}
\title{Convert names into features}
\usage{
as_feature(x, levels)
}
\arguments{
\item{x}{data frame where
name: kanji name
yomi: hiragana}

\item{levels}{a vector of usable kanji&kana characters}
}
\value{
13-dimensional features
}
\description{
Convert names into features
}
