% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderEstimate.R
\name{genderEstimate}
\alias{genderEstimate}
\title{Estimate Gender from kanji and kana of the name}
\usage{
genderEstimate(nm, yomi)
}
\arguments{
\item{nm}{a vector of kanji name}

\item{yomi}{a vector of hiragana name}
}
\value{
a vector of probabilities of the names to be male names
}
\description{
Estimate Gender from kanji and kana of the name
}
